+++ src/EVM/SymExec.hs
@@ -1,11 +1,13 @@
+{-# Language TupleSections #-}
 {-# Language DataKinds #-}
 
 module EVM.SymExec where
 
 import Prelude hiding (Word)
 
+import Data.Tuple (swap)
 import Control.Lens hiding (pre)
-import EVM hiding (Query, Revert, push)
+import EVM hiding (Query, Revert, push, bytecode, cache)
 import qualified EVM
 import EVM.Exec
 import qualified EVM.Fetch as Fetch
@@ -25,8 +27,12 @@ import Control.Concurrent.Async
 import Data.Maybe
 import Data.List (foldl', find)
 import Data.Tuple (swap)
+import Data.Either (isLeft, isRight)
 import Data.ByteString (ByteString)
 import Data.ByteString (null, pack)
+import Data.List (foldl', sortBy)
+import Data.ByteString (ByteString)
+import qualified Data.ByteString as BS
 import qualified Control.Monad.State.Class as State
 import Data.Bifunctor (first, second)
 import Data.Text (Text)
@@ -36,19 +42,26 @@ import qualified Data.Text as T
 import qualified Data.Text.IO as T
 import qualified Data.Text.Lazy as TL
 import qualified Data.Text.Lazy.IO as TL
-import EVM.Format (formatExpr, indent, formatBinary)
+import EVM.Format (formatExpr)
+import Data.Set (Set, isSubsetOf, size)
+import qualified Data.Set as Set
+import Control.Concurrent.STM (atomically, TVar, readTVarIO, readTVar, newTVarIO, writeTVar)
+import Control.Concurrent.Spawn
+import GHC.Conc (getNumProcessors)
+import EVM.Format (indent, formatBinary)
 
 data ProofResult a b c = Qed a | Cex b | SMTTimeout c | SMTError c String
   deriving (Show, Eq)
 type VerifyResult = ProofResult () (Expr End, SMTCex) (Expr End)
-type EquivalenceResult = ProofResult ([VM], [VM]) VM ()
+type EquivResult = ProofResult () (SMTCex) ()
+
+isCex :: ProofResult a b c -> Bool
+isCex (Cex _) = True
+isCex _ = False
 
 isQed :: ProofResult a b c -> Bool
 isQed (Qed _) = True
 isQed _ = False
-
-containsA :: Eq a => Eq b => Eq c => ProofResult a b c -> [(d , e, ProofResult a b c)] -> Bool
-containsA a lst = isJust $ Data.List.find (\(_, _, c) -> c == a) lst
 
 sameCnstr :: ProofResult a b c -> (d , e, ProofResult a b c) -> Bool
 sameCnstr a e = (\(_, _, c) -> checkCnstr c a) e
@@ -241,9 +254,9 @@ interpret fetcher maxIter askSmtIters =
     eval (action Operational.:>>= k) =
       case action of
         Stepper.Exec ->
-          exec >>= interpret fetcher maxIter askSmtIters . k
+          (State.state . runState) exec >>= interpret fetcher maxIter askSmtIters . k
         Stepper.Run ->
-          run >>= interpret fetcher maxIter askSmtIters . k
+          (State.state . runState) run >>= interpret fetcher maxIter askSmtIters . k
         Stepper.IOAct q ->
           mapStateT liftIO q >>= interpret fetcher maxIter askSmtIters . k
         Stepper.Ask (EVM.PleaseChoosePath cond continue) -> do
@@ -288,7 +301,7 @@ maxIterationsReached vm (Just maxIter) =
   let codelocation = getCodeLocation vm
       iters = view (iterations . at codelocation . non 0) vm
   in if num maxIter <= iters
-     then view (cache . path . at (codelocation, iters - 1)) vm
+     then view (EVM.cache . path . at (codelocation, iters - 1)) vm
      else Nothing
 
 
@@ -468,7 +481,7 @@ tooLargeCopy expr = or (foldExpr go [] expr)
   where
     go :: Expr a -> [Bool]
     go = \case
-      (CopySlice _ _ (Lit size) _ _) -> [size > 1024]
+      (CopySlice _ _ (Lit sz) _ _) -> [sz > 1024]
       _  ->  []
 
 getRight ::Either a b -> b
@@ -542,72 +555,144 @@ verify solvers opts preState maybepost = do
         Unsat -> Qed ()
         EVM.SMT.Error e -> SMTError leaf ("SMT Solver Error, solver responded with: " <> show e)
 
+type UnsatCache = TVar [Set Prop]
+
 -- | Compares two contract runtimes for trace equivalence by running two VMs and comparing the end states.
-equivalenceCheck :: SolverGroup -> ByteString -> ByteString -> VeriOpts -> Maybe (Text, [AbiType]) -> IO (Either ExprError [(Maybe SMTCex, Prop, ProofResult () () ())])
+--
+-- We do this by asking the solver to find a common input for each pair of endstates that satisfies the path
+-- conditions for both sides and produces a differing output. If we can find such an input, then we have a clear
+-- equivalence break, and since we run this check for every pair of end states, the check is exhaustive.
+equivalenceCheck :: SolverGroup -> ByteString -> ByteString -> VeriOpts -> Maybe (Text, [AbiType]) -> IO [EquivResult]
 equivalenceCheck solvers bytecodeA bytecodeB opts signature' = do
-  let
-    bytecodeA' = if Data.ByteString.null bytecodeA then Data.ByteString.pack [0] else bytecodeA
-    bytecodeB' = if Data.ByteString.null bytecodeB then Data.ByteString.pack [0] else bytecodeB
-    preStateA = abstractVM signature' [] bytecodeA' Nothing SymbolicS
-    preStateB = abstractVM signature' [] bytecodeB' Nothing SymbolicS
+  case bytecodeA == bytecodeB of
+    True -> do
+      putStrLn "bytecodeA and bytecodeB are identical"
+      pure [Qed ()]
+    False -> do
+      branchesA <- getBranches bytecodeA
+      branchesB <- getBranches bytecodeB
+      let allPairs = [(a,b) | a <- branchesA, b <- branchesB]
+      putStrLn $ "Found " <> (show $ length allPairs) <> " total pairs of endstates"
 
-  aExpr <- evalStateT (interpret (Fetch.oracle solvers Nothing) (maxIter opts) (askSmtIters opts) runExpr) preStateA
-  bExpr <- evalStateT (interpret (Fetch.oracle solvers Nothing) (maxIter opts) (askSmtIters opts) runExpr) preStateB
-  aExprSimp <- if (simp opts) then (pure $ Expr.simplify aExpr) else pure aExpr
-  bExprSimp <- if (simp opts) then (pure $ Expr.simplify bExpr) else pure bExpr
-  when (debug opts) $ putStrLn $ "num of aExprSimp endstates:" <> (show $ length $ flattenExpr aExprSimp) <> "\nnum of bExprSimp endstates:" <> (show $ length $ flattenExpr bExprSimp)
+      when (debug opts) $ putStrLn
+                        $ "endstates in bytecodeA: " <> (show $ length branchesA)
+                       <> "\nendstates in bytecodeB: " <> (show $ length branchesB)
 
-  -- Check each pair of end states for equality:
-  let
-      flattenedA =  flattenExpr aExprSimp
-      flattenedB =  flattenExpr bExprSimp
-      differingEndStates = uncurry distinct <$> [(a,b) | Right a <- flattenedA, Right b <- flattenedB]
-      distinct :: ([Prop], Expr End) -> ([Prop], Expr End) -> Prop
-      distinct (aProps, aEnd) (bProps, bEnd) =
-        let
-          differingResults = case (aEnd, bEnd) of
-            (Return _ aOut aStore, Return _ bOut bStore) ->
-              if aOut == bOut && aStore == bStore then PBool False
-                                                  else aStore ./= bStore  .|| aOut ./= bOut
-            (Return {}, _) -> PBool True
-            (_, Return {}) -> PBool True
-            (Revert _ a, Revert _ b) -> if a==b then PBool False else a ./= b
-            (Revert _ _, _) -> PBool True
-            (_, Revert _ _) -> PBool True
-            (Failure _ erra, Failure _ errb) -> if erra==errb then PBool False else PBool True
-            (GVar _, _) -> error "Expressions cannot contain global vars"
-            (_ , GVar _) -> error "Expressions cannot contain global vars"
-            (Failure _ (WrappedEVMError s), _) -> error $ "Unhandled error: " <> s
-            (_, Failure _ (WrappedEVMError s)) -> error $ "Unhandled error: " <> s
-            (ITE _ _ _, _ ) -> error "Expressions must be flattened"
-            (_, ITE _ _ _) -> error "Expressions must be flattened"
+      let differingEndStates = sortBySize $ mapMaybe (uncurry distinct) allPairs
+      putStrLn $ "Asking the SMT solver for " <> (show $ length differingEndStates) <> " pairs"
+      when (debug opts) $ forM_ (zip differingEndStates [(1::Integer)..]) (\(x, i) ->
+        T.writeFile ("prop-checked-" <> show i) (T.pack $ show x))
 
-        -- if the SMT solver can find a common input that satisfies BOTH sets of path conditions
-        -- AND the output differs, then we are in trouble. We do this for _every_ pair of paths, which
-        -- makes this exhaustive
-        in
-        if differingResults == PBool False
-           then PBool False
-           else (foldl PAnd (PBool True) aProps) .&& (foldl PAnd (PBool True) bProps)  .&& differingResults
-  -- If there exists a pair of end states where this is not the case,
-  -- the following constraint is satisfiable
-  do
-    let diffEndStFilt = filter (/= PBool False) differingEndStates
-    putStrLn $ "Equivalence checking " <> (show $ length diffEndStFilt) <> " combinations"
-    when (debug opts) $ forM_ (zip diffEndStFilt [1 :: Integer ..]) (\(x, i) -> T.writeFile ("prop-checked-" <> show i) (T.pack $ show x))
-    results <- flip mapConcurrently (zip diffEndStFilt [1 :: Integer ..]) $ \(prop, i) -> do
-      let assertedProps = assertProps [prop]
-      let filename = "eq-check-" <> show i <> ".smt2"
-      when (debug opts) $ T.writeFile (filename) $ (TL.toStrict $ formatSMT2 assertedProps) <> "\n(check-sat)"
-      res <- case prop of
-        PBool False -> pure Unsat
-        _ -> checkSat solvers assertedProps
-      case res of
-       Sat a -> return (Just a, prop, Cex ())
-       Unsat -> return (Nothing, prop, Qed ())
-       EVM.SMT.Unknown -> return (Nothing, prop, SMTTimeout ())
-       EVM.SMT.Error txt -> error $ "Error while running solver: `" <> T.unpack txt <> "` SMT file was: `" <> filename <> "`"
-    return $ Right $ filter (\(_, _, res) -> res /= Qed ()) results
+      knownUnsat <- newTVarIO []
+      procs <- getNumProcessors
+      results <- checkAll differingEndStates knownUnsat procs
+
+      let useful = foldr countUseful (0::Integer) results
+      putStrLn $ "Reuse of previous queries was Useful in " <> (show useful) <> " cases"
+      case (all ((==True) . isRight) results) && (all isQed . fmap (fst . getRight) $ results) of
+        True -> pure [Qed ()]
+        False -> pure $ filter (/= Qed ()) . fmap (fst .getRight) $ results ++ filter isLeft results
+  where
+    countUseful :: Either ExprError (EquivResult, Bool) -> Integer -> Integer
+    countUseful (Left _) n = n
+    countUseful (Right (_, b)) n = if b then n+1 else n
+    -- we order the sets by size because this gives us more cache hits when
+    -- running our queries later on (since we rely on a subset check)
+    sortBySize :: [Either ExprError (Set a)] -> [Either ExprError (Set a)]
+    sortBySize = sortBy $ \x y -> case (x,y) of
+                                    ((Left _),_) -> Prelude.GT
+                                    (_, (Left _)) -> Prelude.GT
+                                    ((Right a),(Right b)) -> if size a > size b then Prelude.LT else Prelude.GT
+
+    -- returns True if a is a subset of any of the sets in b
+    subsetAny :: Set Prop -> [Set Prop] -> Bool
+    subsetAny a b = foldr (\bp acc -> acc || isSubsetOf a bp) False b
+
+    -- decompiles the given bytecode into a list of branches
+    getBranches :: ByteString -> IO [Either ExprError ([Prop], Expr End)]
+    getBranches bs = do
+      let
+        bytecode = if BS.null bs then BS.pack [0] else bs
+        prestate = abstractVM signature' [] bytecode Nothing SymbolicS
+      expr <- evalStateT (interpret (Fetch.oracle solvers Nothing) (maxIter opts) (askSmtIters opts) runExpr) prestate
+      let simpl = if (simp opts) then (Expr.simplify expr) else expr
+      pure $ flattenExpr simpl
+
+    -- checks for satisfiability of all the props in the provided set. skips
+    -- the solver if we can determine unsatisfiability from the cache already
+    -- the last element of the returned tuple indicates whether the cache was
+    -- used or not
+    check :: UnsatCache -> Either ExprError (Set Prop) -> IO (Either ExprError (EquivResult, Bool))
+    check knownUnsat props = do
+      if isLeft props then pure $ Left (getLeft props)
+      else do
+        let rightProps = getRight props
+            smt = assertProps $ Set.toList rightProps
+        ku <- readTVarIO knownUnsat
+        res <- if subsetAny rightProps ku
+               then pure (True, Unsat)
+               else (fmap ((False),) (checkSat solvers smt))
+        case res of
+          (_, Sat x) -> pure $ Right (Cex x, False)
+          (quick, Unsat) -> case quick of
+                              True  -> pure $ Right (Qed (), quick)
+                              False -> do
+                                -- nb: we might end up with duplicates here due to a
+                                -- potential race, but it doesn't matter for correctness
+                                atomically $ readTVar knownUnsat >>= writeTVar knownUnsat . (rightProps :)
+                                pure $ Right (Qed (), False)
+          (_, EVM.SMT.Unknown) -> pure $ Right (SMTTimeout (), False)
+          (_, Error txt) -> pure $ Right (SMTError () ("Error while running solver: `" <> show txt), False)
+
+    -- Allows us to run it in parallel. Note that this (seems to) run it
+    -- from left-to-right, and with a max of K threads. This is in contrast to
+    -- mapConcurrently which would spawn as many threads as there are jobs, and
+    -- run them in a random order. We ordered them correctly, though so that'd be bad
+    checkAll :: [Either ExprError (Set Prop)] -> UnsatCache -> Int -> IO [Either ExprError (EquivResult, Bool)]
+    checkAll input cache numproc = do
+       wrap <- pool numproc
+       parMapIO (wrap . (check cache)) input
+
+    -- Takes two branches and returns a set of props that will need to be
+    -- satisfied for the two branches to violate the equivalence check. i.e.
+    -- for a given pair of branches, equivalence is violated if there exists an
+    -- input that satisfies the branch conditions from both sides and produces
+    -- a differing result in each branch
+    distinct :: Either ExprError ([Prop], Expr End) -> Either ExprError ([Prop], Expr End) -> Maybe (Either ExprError (Set Prop))
+    distinct (Left e) _  = Just (Left e)
+    distinct _ (Left e)  = Just (Left e)
+    distinct (Right (aProps, aEnd)) (Right (bProps, bEnd)) =
+      let
+        differingResults = case (aEnd, bEnd) of
+          (Return _ aOut aStore, Return _ bOut bStore) ->
+            if aOut == bOut && aStore == bStore
+            then PBool False
+            else aStore ./= bStore .|| aOut ./= bOut
+          (Return {}, _) -> PBool True
+          (_, Return {}) -> PBool True
+          (Revert _ a, Revert _ b) -> if a == b then PBool False else a ./= b
+          (Revert _ _, _) -> PBool True
+          (_, Revert _ _) -> PBool True
+          (Failure _ (WrappedEVMError s), _) -> error $ "Unhandled error: " <> s
+          (_, Failure _ (WrappedEVMError s)) -> error $ "Unhandled error: " <> s
+          (Failure _ erra, Failure _ errb) -> if erra==errb then PBool False else PBool True
+          (ITE _ _ _, _) -> error "Expressions must be flattened"
+          (_, ITE _ _ _) -> error "Expressions must be flattened"
+          (a, b) -> if a == b
+                    then PBool False
+                    else error $ "Internal Error: Unimplemented. Left: " <> show a <> " Right: " <> show b
+      in case differingResults of
+        -- if the end states are the same, then they can never produce a
+        -- different result under any circumstances
+        PBool False -> Nothing
+        -- if we can statically determine that the end states differ, then we
+        -- ask the solver to find us inputs that satisfy both sets of branch
+        -- conditions
+        PBool True  -> Just . Right . Set.fromList $ aProps <> bProps
+        -- if we cannot statically determine whether or not the end states
+        -- differ, then we ask the solver if the end states can differ if both
+        -- sets of path conditions are satisfiable
+        _ -> Just . Right . Set.fromList $ differingResults : aProps <> bProps
 
 both' :: (a -> b) -> (a, a) -> (b, b)
 both' f (x, y) = (f x, f y)