+++ test/test.hs
@@ -67,6 +67,7 @@ import qualified EVM.Fetch as Fetch
 import Data.List (isSubsequenceOf)
 import Data.Either (isLeft)
 import EVM.TestUtils
+import GHC.Conc (getNumProcessors)
 
 main :: IO ()
 main = defaultMain tests
@@ -182,6 +183,10 @@ tests = testGroup "hevm"
         let simplified = Expr.writeWord idx val buf
             full = WriteWord idx val buf
         checkEquiv simplified full
+    , testProperty "arith-simplification" $ \(_ :: Int) -> ioProperty $ do
+        expr <- generate . sized $ genWordArith 15
+        let simplified = Expr.simplify expr
+        checkEquiv expr simplified
     , testProperty "readByte-equivalance" $ \(buf, idx) -> ioProperty $ do
         let simplified = Expr.readByte idx buf
             full = ReadByte idx buf
@@ -1795,7 +1800,7 @@ tests = testGroup "hevm"
           |]
         withSolvers Z3 3 Nothing $ \s -> do
           a <- equivalenceCheck s aPrgm bPrgm defaultVeriOpts Nothing
-          assertBool "Must have a difference" (not (null a))
+          assertBool "Must have a difference" (any isCex a)
       ,
       testCase "eq-sol-exp-qed" $ do
         Just aPrgm <- solcRuntime "C"
@@ -1820,7 +1825,7 @@ tests = testGroup "hevm"
           |]
         withSolvers Z3 3 Nothing $ \s -> do
           Right a <- equivalenceCheck s aPrgm bPrgm defaultVeriOpts Nothing
-          assertEqual "Must have no difference" [] a
+          assertEqual "Must have no difference" [Qed ()] a
           return ()
       ,
       testCase "eq-sol-exp-cex" $ do
@@ -1848,70 +1853,31 @@ tests = testGroup "hevm"
         withSolvers Z3 3 Nothing $ \s -> do
           let myVeriOpts = VeriOpts{ simp = True, debug = False, maxIter = Just 2, askSmtIters = Just 2, rpcInfo = Nothing}
           Right a <- equivalenceCheck s aPrgm bPrgm myVeriOpts Nothing
-          assertEqual "Must be different" (containsA (Cex ()) a) True
+          assertEqual "Must be different" (any isCex a) True
           return ()
       , testCase "eq-all-yul-optimization-tests" $ do
         let myVeriOpts = VeriOpts{ simp = True, debug = False, maxIter = Just 5, askSmtIters = Just 20, rpcInfo = Nothing }
             ignoredTests = [
-                      "controlFlowSimplifier/terminating_for_nested.yul"
-                    , "controlFlowSimplifier/terminating_for_nested_reversed.yul"
-
                     -- unbounded loop --
-                    , "commonSubexpressionEliminator/branches_for.yul"
-                    , "commonSubexpressionEliminator/loop.yul"
-                    , "conditionalSimplifier/clear_after_if_continue.yul"
+                    "commonSubexpressionEliminator/branches_for.yul"
                     , "conditionalSimplifier/no_opt_if_break_is_not_last.yul"
-                    , "conditionalUnsimplifier/clear_after_if_continue.yul"
                     , "conditionalUnsimplifier/no_opt_if_break_is_not_last.yul"
                     , "expressionSimplifier/inside_for.yul"
                     , "forLoopConditionIntoBody/cond_types.yul"
                     , "forLoopConditionIntoBody/simple.yul"
                     , "fullSimplify/inside_for.yul"
-                    , "fullSuite/devcon_example.yul"
-                    , "fullSuite/loopInvariantCodeMotion.yul"
                     , "fullSuite/no_move_loop_orig.yul"
-                    , "loadResolver/loop.yul"
                     , "loopInvariantCodeMotion/multi.yul"
-                    , "loopInvariantCodeMotion/recursive.yul"
-                    , "loopInvariantCodeMotion/simple.yul"
-                    , "redundantAssignEliminator/for_branch.yul"
-                    , "redundantAssignEliminator/for_break.yul"
-                    , "redundantAssignEliminator/for_continue.yul"
-                    , "redundantAssignEliminator/for_decl_inside_break_continue.yul"
-                    , "redundantAssignEliminator/for_deep_noremove.yul"
                     , "redundantAssignEliminator/for_deep_simple.yul"
-                    , "redundantAssignEliminator/for_multi_break.yul"
-                    , "redundantAssignEliminator/for_nested.yul"
-                    , "redundantAssignEliminator/for_rerun.yul"
-                    , "redundantAssignEliminator/for_stmnts_after_break_continue.yul"
-                    , "rematerialiser/branches_for1.yul"
-                    , "rematerialiser/branches_for2.yul"
-                    , "rematerialiser/for_break.yul"
-                    , "rematerialiser/for_continue.yul"
-                    , "rematerialiser/for_continue_2.yul"
-                    , "rematerialiser/for_continue_with_assignment_in_post.yul"
-                    , "rematerialiser/no_remat_in_loop.yul"
-                    , "ssaTransform/for_reassign_body.yul"
-                    , "ssaTransform/for_reassign_init.yul"
-                    , "ssaTransform/for_reassign_post.yul"
-                    , "ssaTransform/for_simple.yul"
-                    , "loopInvariantCodeMotion/nonMovable.yul"
-                    , "unusedAssignEliminator/for_rerun.yul"
-                    , "unusedAssignEliminator/for_continue_3.yul"
+                    , "unusedAssignEliminator/for_deep_noremove.yul"
                     , "unusedAssignEliminator/for_deep_simple.yul"
                     , "ssaTransform/for_def_in_init.yul"
-                    , "rematerialiser/many_refs_small_cost_loop.yul"
+                    , "loopInvariantCodeMotion/simple_state.yul"
+                    , "loopInvariantCodeMotion/simple.yul"
+                    , "loopInvariantCodeMotion/recursive.yul"
+                    , "loopInvariantCodeMotion/no_move_staticall_returndatasize.yul"
                     , "loopInvariantCodeMotion/no_move_state_loop.yul"
-                    , "loopInvariantCodeMotion/dependOnVarInLoop.yul"
-                    , "forLoopInitRewriter/empty_pre.yul"
-                    , "loadResolver/keccak_crash.yul"
-                    , "blockFlattener/for_stmt.yul" -- symb input can loop it forever
-                    , "unusedAssignEliminator/for.yul" -- not infinite, just 2**256-3
                     , "loopInvariantCodeMotion/no_move_state.yul" -- not infinite, but rollaround on a large int
-                    , "loopInvariantCodeMotion/non-ssavar.yul" -- same as above
-                    , "forLoopInitRewriter/complex_pre.yul"
-                    , "rematerialiser/some_refs_small_cost_loop.yul" -- not infinite but 100 long
-                    , "forLoopInitRewriter/simple.yul"
                     , "loopInvariantCodeMotion/no_move_loop.yul"
 
                     -- unexpected symbolic arg --
@@ -1961,11 +1927,15 @@ tests = testGroup "hevm"
                     , "fullSuite/ssaReverse.yul"
                     , "rematerialiser/cheap_caller.yul"
                     , "rematerialiser/non_movable_instruction.yul"
+                    , "rematerialiser/for_break.yul"
+                    , "rematerialiser/for_continue.yul"
+                    , "rematerialiser/for_continue_2.yul"
                     , "ssaAndBack/multi_assign.yul"
                     , "ssaAndBack/multi_assign_if.yul"
                     , "ssaAndBack/multi_assign_switch.yul"
                     , "ssaAndBack/simple.yul"
                     , "ssaReverser/simple.yul"
+                    , "loopInvariantCodeMotion/simple_storage.yul"
 
                     -- OpMstore8
                     , "loadResolver/memory_with_different_kinds_of_invalidation.yul"
@@ -1980,7 +1950,6 @@ tests = testGroup "hevm"
                     , "commonSubexpressionEliminator/object_access.yul"
                     , "expressionSplitter/object_access.yul"
                     , "fullSuite/stack_compressor_msize.yul"
-                    , "varNameCleaner/function_names.yul"
 
                     -- stack too deep --
                     , "fullSuite/abi2.yul"
@@ -2080,8 +2049,6 @@ tests = testGroup "hevm"
 
                     -- Takes too long, would timeout on most test setups.
                     -- We could probably fix these by "bunching together" queries
-                    , "fullSuite/clear_after_if_continue.yul"
-                    , "reasoningBasedSimplifier/smod.yul"
                     , "reasoningBasedSimplifier/mulmod.yul"
 
                     -- TODO check what's wrong with these!
@@ -2109,9 +2076,8 @@ tests = testGroup "hevm"
                 False -> recursiveList ax (a:b)
           recursiveList [] b = pure b
         files <- recursiveList fullpaths []
-        --
         let filesFiltered = filter (\file -> not $ any (\filt -> Data.List.isSubsequenceOf filt file) ignoredTests) files
-        --
+
         -- Takes one file which follows the Solidity Yul optimizer unit tests format,
         -- extracts both the nonoptimized and the optimized versions, and checks equivalence.
         forM_ filesFiltered (\f-> do
@@ -2149,7 +2115,9 @@ tests = testGroup "hevm"
             putStrLn "------------- END -----------------"
           Just aPrgm <- yul "" $ T.pack $ unlines filteredASym
           Just bPrgm <- yul "" $ T.pack $ unlines filteredBSym
-          withSolvers CVC5 6 (Just 3) $ \s -> do
+          procs <- getNumProcessors
+          withSolvers CVC5 (num procs) (Just 100) $ \s -> do
+            res <- equivalenceCheck s aPrgm bPrgm myVeriOpts Nothing
             Right res <- equivalenceCheck s aPrgm bPrgm myVeriOpts Nothing
             end <- getCurrentTime
             let cexs = filter (sameCnstr (Cex())) res
@@ -2698,6 +2666,46 @@ genWord litFreq sz = frequency
    subBuf = defaultBuf (sz `div` 10)
    subStore = genStorage (sz `div` 10)
    subByte = genByte (sz `div` 10)
+
+genWordArith :: Int -> Int -> Gen (Expr EWord)
+genWordArith litFreq 0 = frequency
+  [ (litFreq, fmap Lit arbitrary)
+  , (1, oneof [ fmap Lit arbitrary ])
+  ]
+genWordArith litFreq sz = frequency
+  [ (litFreq, fmap Lit arbitrary)
+  , (20, frequency
+    [ (20, liftM2 Add  subWord subWord)
+    , (20, liftM2 Sub  subWord subWord)
+    , (20, liftM2 Mul  subWord subWord)
+    , (20, liftM2 SEx  subWord subWord)
+    , (20, liftM2 Xor  subWord subWord)
+    -- these reduce variability
+    , (3 , liftM2 Min  subWord subWord)
+    , (3 , liftM2 Div  subWord subWord)
+    , (3 , liftM2 SDiv subWord subWord)
+    , (3 , liftM2 Mod  subWord subWord)
+    , (3 , liftM2 SMod subWord subWord)
+    , (3 , liftM2 SHL  subWord subWord)
+    , (3 , liftM2 SHR  subWord subWord)
+    , (3 , liftM2 SAR  subWord subWord)
+    , (3 , liftM2 Or   subWord subWord)
+    -- comparisons, reducing variability greatly
+    , (1 , liftM2 LEq  subWord subWord)
+    , (1 , liftM2 GEq  subWord subWord)
+    , (1 , liftM2 SLT  subWord subWord)
+    --(1, , liftM2 SGT subWord subWord
+    , (1 , liftM2 Eq   subWord subWord)
+    , (1 , liftM2 And  subWord subWord)
+    , (1 , fmap IsZero subWord        )
+    -- Expensive below
+    --(1,  liftM3 AddMod subWord subWord subWord
+    --(1,  liftM3 MulMod subWord subWord subWord
+    --(1,  liftM2 Exp subWord litWord
+    ])
+  ]
+ where
+   subWord = genWordArith (litFreq `div` 2) (sz `div` 2)
 
 defaultBuf :: Int -> Gen (Expr Buf)
 defaultBuf = genBuf (4_000_000)